chore: 低危卫生批量修复（#77）
- SQL LIMIT 由常量插值生成，消除与导出常量的静默漂移风险
- mergeModelUsage / mergeUsage 双实现收敛到 model-stream.types.ts 单一导出
- AgentMarkdownContent 的 MarkdownIt 实例提升为模块级共享
- web http.ts isApiSuccessResponse 补齐 success === true 校验（对齐 admin）
- 删除 admin 无引用的 defaultRunFilters 与随之无用的类型导入
- 更新 web api/seo.ts 过时注释；删除 runtime 无信息量注释
