+++ apps/api/src/retrieval/postgres-article-retrieval.repository.ts
@@ -148,7 +148,7 @@ const LEXICAL_CANDIDATES_SQL = String.raw`
       ELSE 4
     END ASC,
     article."sourceId" ASC
-  LIMIT 10
+  LIMIT ${LEXICAL_ARTICLE_CANDIDATE_LIMIT}
 `
 
 const VECTOR_CANDIDATES_SQL = `
@@ -211,7 +211,7 @@ const VECTOR_CANDIDATES_SQL = `
     source_id ASC,
     ordinal ASC,
     chunk_id ASC
-  LIMIT 40
+  LIMIT ${VECTOR_CHUNK_CANDIDATE_LIMIT}
 `
 
 const SET_LOCAL_TIMEOUTS_SQL = `
